Fix saveToFile

+++ index.js
@@ -86,7 +86,10 @@ function saveToFile(
   let fileContent = content;
   if (fs.existsSync(fullPath)) {
     fileContent = JSON.parse(fs.readFileSync(fullPath, 'utf-8'));
-    fs.writeFileSync(fullPath.replace('.json', '-copy.json'), stringify(fileContent));
+    fs.writeFileSync(
+      fullPath.replace('.json', '-copy.json'),
+      stringify(fileContent)
+    );
     if (!key) {
       fileContent instanceof Array
         ? (fileContent = [...fileContent, content])
@@ -99,7 +102,9 @@ function saveToFile(
   }
   fileContent = stringify(fileContent);
   fs.writeFileSync(fullPath, fileContent);
-  fsp.unlink(fullPath.replace('.json', '-copy.json'));
+  if (fs.existsSync(fullPath.replace('.json', '-copy.json'))) {
+    fsp.unlink(fullPath.replace('.json', '-copy.json'));
+  }
 }
 
 /**
